Updated dialog to common module
